telegram_bot/requirements.txt
# python==3.7.4
python-dotenv
aiogram
openai==0.27.0
